Add files via upload
The tool is just for explanation, the operating system can have this feature in built or can have a short cut key for it. In order to use the tool open the .exe file and click on turn on  now you are ready to use ctrl+c on various resources multiple times , remember to copy in the order you want to paste the texts. Once you have copied all the texts, finally use ctrl+v to paste all the texts. Just turn off and then on to refresh the tool.